refactor: split transaction module into mod.rs and update.rs
Reorganize the transaction module so that each file has a clear responsibility:

- mod.rs: shared Transaction struct, fields, commit logic, and common helpers
- create_table.rs: CreateTableTransaction type alias and try_new_create_table constructor
- update.rs (new): ExistingTableTransaction type alias, try_new_existing_table constructor,
  existing-table-only public API (with_blind_append, with_operation,
  with_domain_metadata_removed, with_transaction_id, remove_files, with_committer),
  deletion vector update logic (update_deletion_vectors, DvMatchVisitor), and all
  DV-related schemas/statics

Also:
- Remove with_committer from the shared impl; it now lives in update.rs on
  ExistingTableTransaction only
- Update the catalog-managed table error message to reference Snapshot::transaction()
  instead of Transaction::with_committer()
- Update the corresponding assertion in filesystem.rs tests

This is a pure refactor with no behavioral changes.
